requirements.txt
streamlit
pandas
python-docx
PyPDF2
langchain
python-dotenv
langchain-community
langchain_openai